Adds asset manager
Allows for creating a database of images with corner points tagged in
a single file. This persistance allows us to add ordinary images to
our database so we can run experiments on them.

Next goal is to make this interactive with OpenCV so that the corners
can be easily clicked, making the whiteboard tagging rapid, instead of
having to look up the corners through an image editor like GIMP.