Makes adding commands easier
Adds the idea of cmdlets, which are basically just arg parsers but
organized specifically for specific commands. Since this tool is
command oriented it helps to have an object that makes building and
parameterizing these simple. We have all the flexibility that
argparse provides, but this is a bit more ergonomic.