Moves image manipulation into a dedicated class
Creates an image class that encapsulates image manipulations and
dealing with opencv directly. Now we can just make an image and apply
some transformations through a much, much better interface, with less
parameters and probably much more safetey as well. This implements
most of the functionality that was in our test file, but there's still
more to be done, and of course tested.