Adds ability to save, improves initializer
When we attempt to read or write a file using imread/write, we first
attempt to get a file handle and then immediately close it. This is
becaue the imread/write functions will not raise an error, but instead
just return an empty array. By getting the file handle ahead of time,
we force Python to throw the regular file errors we're used to, and
can program aound those constructs.